Drug resistance muts in separate script
